app/requirements.txt
asgiref==3.8.1
Django==5.1.2
pillow==11.0.0
sqlparse==0.5.1
## The following requirements were added by pip freeze:
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-texttospeech==2.21.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
idna==3.10
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
urllib3==2.2.3